* release/20161221: (192 commits)
  CommentManager::getMostVotedCommentsforContentID return only value_pos comments
  Avoid errors for RSS not valid in News Agency
  Fix Newsletter subscriber batch delete controller action
  Fix newsletter search by title in backend
  Fix thumb visualization for image listing
  Add opinion link in backend form
  Fix layout offline message for backend
  Add Adblock detector in backend advertisement module
  Use new persistor for Backend author save action
  Go back to page 1 when changing the epp on content listings
  ONM-2041: Fix undefined extension in domains
  Fix accordion ids in system settings form
  Avoid crashes for unsuccessfull ftp connections in news agency
  Fix mispelled "sucess" string
  Fix variable name in backend web service subscribers
  Fix in widget checker for special backend listing
  ONM-2048: Load instance from OQL given its database name
  Fix wrong column order in special list backen
  Fix template condition check for category in_rss
  ONM-2041: Remove payment details in invoice preview
  ...

+++ src/BackendWebService/Controller/ContentController.php
@@ -11,7 +11,7 @@
 namespace BackendWebService\Controller;
 
 use Common\Core\Annotation\Security;
-use Onm\Framework\Controller\Controller;
+use Common\Core\Controller\Controller;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -108,12 +108,21 @@ class ContentController extends Controller
 
         if (!is_null($content->id)) {
             try {
-                $content->delete($id);
-                $success[] = array(
-                    'id'      => $id,
-                    'message' => _('Item deleted successfully'),
-                    'type'    => 'success'
-                );
+                if (!$this->get('core.security')->hasPermission('CONTENT_OTHER_UPDATE')
+                    && !$content->isOwner($this->getUser()->id)
+                ) {
+                    $errors[] = [
+                        'error',
+                        _("You don't have enough privileges for modify this content.")
+                    ];
+                } else {
+                    $content->delete($id);
+                    $success[] = array(
+                        'id'      => $id,
+                        'message' => _('Item deleted successfully'),
+                        'type'    => 'success'
+                    );
+                }
             } catch (Exception $e) {
                 $errors[] = array(
                     'id'      => $id,
@@ -160,8 +169,20 @@ class ContentController extends Controller
 
                 if (!is_null($content->id)) {
                     try {
-                        $content->delete($id);
-                        $updated[] = $id;
+                        if (!$this->get('core.security')->hasPermission('CONTENT_OTHER_UPDATE')
+                            && !$content->isOwner($this->getUser()->id)
+                        ) {
+                            $errors[] = [
+                                'error',
+                                sprintf(
+                                    _("You don't have enough privileges to delete the content with id %s."),
+                                    $content->id
+                                )
+                            ];
+                        } else {
+                            $content->delete($id);
+                            $updated[] = $id;
+                        }
                     } catch (Exception $e) {
                         $errors[] = array(
                             'id'      => $id,
